Add TextRun height

+++ nfm-world/driverinterface/IFontMetrics.cs
@@ -3,5 +3,5 @@ namespace NFMWorld.DriverInterface;
 public interface IFontMetrics
 {
     public float StringWidth(string astring);
-    public int Height(string astring);
+    public float Height(string astring);
 }
